Remove aes sub-package from exercises
