--- src/cryptoserve/exercises/primitives/__init__.py
@@ -1,1 +1,0 @@
-from cryptoserve.exercises.primitives.aes import *
